[front/components/data_source] - feature: add tabs for documents and tables in data source view
 - Implement separate tab views for managing documents and tables associated with a data source
 - Enable navigation and functionality for adding new documents or tables through UI interaction

[front/components/data_source] - refactor: extract documents and tables tab views to separate components

 - Improve maintainability by dividing the standard data source view into smaller, more focused components
 - Enhance code organization reflecting the separation of concerns for documents and tables management

[front] - fix: correct write permission check for data sources

 - Update permission check to allow users with 'write' permission or 'builder' status to modify data sources
 - Resolve an issue where the validation logic incorrectly restricted write permissions to builders only
